US5238 - Camp Thank You (#3144)
* US5378 getting organized

* US5378 Adding path for payments (non-donation)

* US5378 WIP

* US5238 - Create framework and route for Thank You camp page; Extract the family members into a sub-component that can be used on thank you page

* US5238 - Basic test for camps thank you page

* US5378 clean up TODOs and hard coded vars

* US5238 - dynamically set camp heading for thank you page; Re-arrange text on page; Add content block for signing up other students

* US5378 clean up exisiting tests

* US5378 Added metadata to stripe charge for payment

* Modify paymentApi to return generated ids

Added tests

* US5378 Cleanup

* Dynamic Resolves!

* US5378 remove comment

* US5378 new payment

* US5238

Start new tests

* US5378 Change double to decimal and refactor

* US5378 test fix

* US5238 fix compile errors

* US5238 - Thank you page calls payment API and populates amount paid, amount remaining, and email address for receipt

* US5238 - Thank you page tests

* US5238 - Fix issue with medical form test; Allow for camp specific thank you content block

* US5238 - Fix broken payment API test

+++ Gateway/MinistryPlatform.Translation/Models/Payments/MpNestedInvoiceDetail.cs
@@ -22,7 +22,7 @@ namespace MinistryPlatform.Translation.Models.Payments
         public int Quantity { get; set; }
 
         [JsonProperty(PropertyName = "Line_Total")]
-        public double LineTotal { get; set; }
+        public decimal LineTotal { get; set; }
 
         [JsonProperty(PropertyName = "Product_ID")]
         public int ProductId { get; set; }
