+++ Gateway/crds-angular/Models/Crossroads/Camp/ProductOptionDTO.cs
@@ -12,13 +12,13 @@ namespace crds_angular.Models.Crossroads.Camp
         public string OptionTitle { get; set; }
 
         [JsonProperty(PropertyName = "optionPrice")]
-        public double OptionPrice { get; set; }
+        public decimal OptionPrice { get; set; }
 
         [JsonProperty(PropertyName = "daysOutToHide")]
         public int? DaysOutToHide { get; set; }
 
         [JsonProperty(PropertyName = "totalWithOptionPrice")]
-        public double TotalWithOptionPrice { get; set; }
+        public decimal TotalWithOptionPrice { get; set; }
 
         [JsonProperty(PropertyName = "endDate")]
         public DateTime? EndDate { get; set; }
